# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface VehicleInput {
  plate: string;
  brand: string;
  model: string;
  year: number;
  currentKm: number;
  lastMaintenanceKm: number;
  maintenanceInterval: number;
  nextMaintenanceKm: number;
  active: boolean;
}

export interface DriverInput {
  name: string;
  pin?: string;
  active: boolean;
}

export interface TripInput {
  driverId: number;
  vehicleId: number;
  initialKm: number;
  finalKm: number;
  origin: string;
  destination: string;
  companions?: string;
  refueled: boolean;
  liters?: number;
  fuelCost?: number;
  observations?: string;
}

export interface MaintenanceInput {
  vehicleId: number;
  date: string;
  km: number;
  type: string;
  description: string;
  cost: number;
  observations: string;
  nextMaintenanceKm: number;
}

export type LoginInputRole = typeof LoginInputRole[keyof typeof LoginInputRole];


export const LoginInputRole = {
  admin: 'admin',
  driver: 'driver',
} as const;

export interface LoginInput {
  role: LoginInputRole;
  driverId?: number;
  pin: string;
}

export interface SetupInput {
  name: string;
  pin: string;
  email: string;
}

export interface SettingsInput {
  adminEmail: string;
  senderEmail: string;
}

export interface CorrectionInput {
  trip: TripInput;
  reason: string;
}

export interface Session {
  authenticated: boolean;
  setupRequired: boolean;
  role?: string;
  name?: string;
  driverId?: number;
}

export interface Result {
  message: string;
}

export interface Driver {
  id: number;
  name: string;
  active: boolean;
}

export interface Audit {
  id: number;
  at: string;
  actor: string;
  action: string;
  reason: string;
  before: string;
  after: string;
}

export interface FuelSummary {
  vehicleId: number;
  plate: string;
  liters: number;
  cost: number;
  loads: number;
  averageLiters: number;
  distance: number;
  kmPerLiter: number;
}

export interface Settings {
  adminEmail: string;
  senderEmail: string;
  emailConnected: boolean;
  pendingAlerts: number;
}

export type VehicleStatus = typeof VehicleStatus[keyof typeof VehicleStatus];


export const VehicleStatus = {
  normal: 'normal',
  warning: 'warning',
  urgent: 'urgent',
  overdue: 'overdue',
} as const;

export interface Vehicle {
  plate: string;
  brand: string;
  model: string;
  year: number;
  currentKm: number;
  lastMaintenanceKm: number;
  maintenanceInterval: number;
  nextMaintenanceKm: number;
  active: boolean;
  id: number;
  status: VehicleStatus;
  remainingKm: number;
  lastDriver: string;
  lastRecord: string;
}

export interface Trip {
  driverId: number;
  vehicleId: number;
  initialKm: number;
  finalKm: number;
  origin: string;
  destination: string;
  companions: string;
  refueled: boolean;
  liters: number;
  fuelCost: number;
  observations: string;
  id: number;
  at: string;
  driverName: string;
  plate: string;
  distance: number;
  corrected: boolean;
}

export interface Maintenance {
  vehicleId: number;
  date: string;
  km: number;
  type: string;
  description: string;
  cost: number;
  observations: string;
  nextMaintenanceKm: number;
  id: number;
  plate: string;
  actor: string;
}

export interface Dashboard {
  vehicles: Vehicle[];
  recentTrips: Trip[];
  totalTrips: number;
  totalDistance: number;
  fuelCost: number;
  activeDrivers: number;
}

export type ListTripsParams = {
from?: string;
to?: string;
vehicleId?: string;
driverId?: string;
refueled?: string;
plate?: string;
};

export type GetFuelSummaryParams = {
from?: string;
to?: string;
vehicleId?: string;
driverId?: string;
refueled?: string;
plate?: string;
};