# lib/api-spec/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Api",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/setup": {
      "post": {
        "operationId": "setupAdmin",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupInput"
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "operationId": "login",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        }
      }
    },
...
    "/vehicles": {
      "get": {
        "operationId": "listVehicles",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Vehicle"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createVehicle",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleInput"
              }
            }
          }
        }
      }
    },
    "/drivers": {
      "get": {
        "operationId": "listDrivers",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Driver"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createDriver",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Driver"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverInput"
              }
            }
          }
        }
      }
    },
...
    "/maintenances": {
      "get": {
        "operationId": "listMaintenances",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Maintenance"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createMaintenance",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Maintenance"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaintenanceInput"
              }
            }
          }
        }
      }
    },
    "/vehicles/{id}": {
      "put": {
        "operationId": "updateVehicle",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vehicle"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VehicleInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/drivers/{id}": {
      "put": {
        "operationId": "updateDriver",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Driver"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DriverInput"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
...
    "/settings": {
      "get": {
        "operationId": "getSettings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateSettings",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Settings"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsInput"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "VehicleInput": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "currentKm": {
            "type": "number"
          },
          "lastMaintenanceKm": {
            "type": "number"
          },
          "maintenanceInterval": {
            "type": "number"
          },
          "nextMaintenanceKm": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "plate",
          "brand",
          "model",
          "year",
          "currentKm",
          "lastMaintenanceKm",
          "maintenanceInterval",
          "nextMaintenanceKm",
          "active"
        ]
      },
      "DriverInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "pin": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "name",
          "active"
        ]
      },
...
      "MaintenanceInput": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "integer"
          },
          "date": {
            "type": "string"
          },
          "km": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "cost": {
            "type": "number"
          },
          "observations": {
            "type": "string"
          },
          "nextMaintenanceKm": {
            "type": "number"
          }
        },
        "required": [
          "vehicleId",
          "date",
          "km",
          "type",
          "description",
          "cost",
          "observations",
          "nextMaintenanceKm"
        ]
      },
      "LoginInput": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "driver"
            ]
          },
          "driverId": {
            "type": "integer"
          },
          "pin": {
            "type": "string"
          }
        },
        "required": [
          "role",
          "pin"
        ]
      },
      "SetupInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "pin": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "pin",
          "email"
        ]
      },
      "SettingsInput": {
        "type": "object",
        "properties": {
          "adminEmail": {
            "type": "string"
          },
          "senderEmail": {
            "type": "string"
          }
        },
        "required": [
          "adminEmail",
          "senderEmail"
        ]
      },
...
      "Session": {
        "type": "object",
        "properties": {
          "authenticated": {
            "type": "boolean"
          },
          "setupRequired": {
            "type": "boolean"
          },
          "role": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "driverId": {
            "type": "integer"
          }
        },
        "required": [
          "authenticated",
          "setupRequired"
        ]
      },
...
      "Driver": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "name",
          "active"
        ]
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "adminEmail": {
            "type": "string"
          },
          "senderEmail": {
            "type": "string"
          },
          "emailConnected": {
            "type": "boolean"
          },
          "pendingAlerts": {
            "type": "integer"
          }
        },
        "required": [
          "adminEmail",
          "senderEmail",
          "emailConnected",
          "pendingAlerts"
        ]
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "plate": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "year": {
            "type": "integer"
          },
          "currentKm": {
            "type": "number"
          },
          "lastMaintenanceKm": {
            "type": "number"
          },
          "maintenanceInterval": {
            "type": "number"
          },
          "nextMaintenanceKm": {
            "type": "number"
          },
          "active": {
            "type": "boolean"
          },
          "id": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "normal",
              "warning",
              "urgent",
              "overdue"
            ]
          },
          "remainingKm": {
            "type": "number"
          },
          "lastDriver": {
            "type": "string"
          },
          "lastRecord": {
            "type": "string"
          }
        },
        "required": [
          "plate",
          "brand",
          "model",
          "year",
          "currentKm",
          "lastMaintenanceKm",
          "maintenanceInterval",
          "nextMaintenanceKm",
          "active",
          "id",
          "status",
          "remainingKm",
          "lastDriver",
          "lastRecord"
        ]
      },
...
      "Maintenance": {
        "type": "object",
        "properties": {
          "vehicleId": {
            "type": "integer"
          },
          "date": {
            "type": "string"
          },
          "km": {
            "type": "number"
          },
          "type": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "cost": {
            "type": "number"
          },
          "observations": {
            "type": "string"
          },
          "nextMaintenanceKm": {
            "type": "number"
          },
          "id": {
            "type": "integer"
          },
          "plate": {
            "type": "string"
          },
          "actor": {
            "type": "string"
          }
        },
        "required": [
          "vehicleId",
          "date",
          "km",
          "type",
          "description",
          "cost",
          "observations",
          "nextMaintenanceKm",
          "id",
          "plate",
          "actor"
        ]
      },
...
    }
  }
}