# lib/api-zod/src/generated/types/maintenance.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Maintenance {
  vehicleId: number;
  date: string;
  km: number;
  type: string;
  description: string;
  cost: number;
  observations: string;
  nextMaintenanceKm: number;
  id: number;
  plate: string;
  actor: string;
}