# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface Settings {
  adminEmail: string;
  senderEmail: string;
  emailConnected: boolean;
  pendingAlerts: number;
}