# lib/api-zod/src/generated/types/vehicle.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { VehicleStatus } from './vehicleStatus';

export interface Vehicle {
  plate: string;
  brand: string;
  model: string;
  year: number;
  currentKm: number;
  lastMaintenanceKm: number;
  maintenanceInterval: number;
  nextMaintenanceKm: number;
  active: boolean;
  id: number;
  status: VehicleStatus;
  remainingKm: number;
  lastDriver: string;
  lastRecord: string;
}