# lib/api-zod/src/generated/types/vehicleInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface VehicleInput {
  plate: string;
  brand: string;
  model: string;
  year: number;
  currentKm: number;
  lastMaintenanceKm: number;
  maintenanceInterval: number;
  nextMaintenanceKm: number;
  active: boolean;
}